backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
sqlalchemy==2.0.23
python-multipart==0.0.6
